+++ fakeredis/_basefakesocket.py
@@ -8,10 +8,10 @@ from xmlrpc.client import ResponseError
 import redis
 from redis.connection import DefaultParser
 
-from fakeredis.model import XStream, ZSet
+from fakeredis.model import XStream, ZSet, Hash, ExpiringMembersSet
 from . import _msgs as msgs
 from ._command_args_parsing import extract_args
-from ._commands import Int, Float, SUPPORTED_COMMANDS, COMMANDS_WITH_SUB, Signature, CommandItem, Hash
+from ._commands import Int, Float, SUPPORTED_COMMANDS, COMMANDS_WITH_SUB, Signature, CommandItem
 from ._helpers import (
     SimpleError,
     valid_response_type,
@@ -394,7 +394,7 @@ class BaseFakeSocket:
             return SimpleString(b"string")
         elif isinstance(key.value, list):
             return SimpleString(b"list")
-        elif isinstance(key.value, set):
+        elif isinstance(key.value, ExpiringMembersSet):
             return SimpleString(b"set")
         elif isinstance(key.value, ZSet):
             return SimpleString(b"zset")