+++ fakeredis/commands_mixins/generic_mixin.py
@@ -14,10 +14,9 @@ from fakeredis._commands import (
     CommandItem,
     SortFloat,
     delete_keys,
-    Hash,
 )
 from fakeredis._helpers import compile_pattern, SimpleError, OK, casematch, Database, SimpleString
-from fakeredis.model import ZSet
+from fakeredis.model import ZSet, Hash, ExpiringMembersSet
 
 
 class GenericCommandsMixin:
@@ -224,7 +223,7 @@ class GenericCommandsMixin:
 
     @command(name="SORT", fixed=(Key(),), repeat=(bytes,))
     def sort(self, key, *args):
-        if key.value is not None and not isinstance(key.value, (set, list, ZSet)):
+        if key.value is not None and not isinstance(key.value, (ExpiringMembersSet, list, ZSet)):
             raise SimpleError(msgs.WRONGTYPE_MSG)
         (
             asc,