+++ docs/about/changelog.md
@@ -13,6 +13,7 @@ toc_depth: 2
 
 - ACL commands support #338
 - Add support disable_decoding in async read_response #349
+- Implement support for `SADDEX`, using a new set implementation with support for expiring members #350
 
 
 ### 🧰 Maintenance